WebAuthorizationActivityResultDispatcher and related refactoring in 3DS and APM launchers
