+++ sdk/src/main/kotlin/com/processout/sdk/ui/threeds/PO3DSRedirectCustomTabLauncher.kt
@@ -8,13 +8,11 @@ import androidx.fragment.app.Fragment
 import com.processout.sdk.api.ProcessOut
 import com.processout.sdk.api.model.threeds.PO3DSRedirect
 import com.processout.sdk.api.network.ApiConstants
-import com.processout.sdk.core.POFailure
-import com.processout.sdk.core.ProcessOutActivityResult
-import com.processout.sdk.core.ProcessOutResult
+import com.processout.sdk.core.*
 import com.processout.sdk.core.logger.POLogger
+import com.processout.sdk.ui.web.WebAuthorizationActivityResultDispatcher
 import com.processout.sdk.ui.web.WebAuthorizationDelegate
 import com.processout.sdk.ui.web.WebAuthorizationDelegateCache
-import com.processout.sdk.ui.web.WebAuthorizationDelegateMemoryCache
 import com.processout.sdk.ui.web.customtab.CustomTabAuthorizationActivityContract
 import com.processout.sdk.ui.web.customtab.CustomTabConfiguration
 import com.processout.sdk.ui.web.webview.WebViewAuthorizationActivityLauncher
@@ -38,7 +36,7 @@ class PO3DSRedirectCustomTabLauncher private constructor(
          * __Note:__ Required to call in _onCreate()_ to register for activity result.
          */
         fun create(from: Fragment) = PO3DSRedirectCustomTabLauncher(
-            WebAuthorizationDelegateMemoryCache
+            WebAuthorizationActivityResultDispatcher
         ).apply {
             customTabLauncher = from.registerForActivityResult(
                 CustomTabAuthorizationActivityContract(),
@@ -54,7 +52,7 @@ class PO3DSRedirectCustomTabLauncher private constructor(
          * __Note:__ Required to call in _onCreate()_ to register for activity result.
          */
         fun create(from: ComponentActivity) = PO3DSRedirectCustomTabLauncher(
-            WebAuthorizationDelegateMemoryCache
+            WebAuthorizationActivityResultDispatcher
         ).apply {
             customTabLauncher = from.registerForActivityResult(
                 CustomTabAuthorizationActivityContract(),
@@ -126,15 +124,13 @@ class PO3DSRedirectCustomTabLauncher private constructor(
         launch(redirect, returnUrl = String(), callback)
     }
 
-    private val activityResultCallback = ActivityResultCallback<ProcessOutActivityResult<Uri>> {
-        if (delegateCache.isCached().not()) {
-            POLogger.error("Cannot provide 3DS result. Delegate is not cached.")
+    private val activityResultCallback = ActivityResultCallback<ProcessOutActivityResult<Uri>> { result ->
+        if (!delegateCache.isCached()) {
+            POLogger.error("Cannot provide result. Delegate is not cached. Possibly process was killed.")
+            return@ActivityResultCallback
         }
-        when (it) {
-            is ProcessOutActivityResult.Success -> delegateCache.remove()?.complete(uri = it.value)
-            is ProcessOutActivityResult.Failure -> delegateCache.remove()?.complete(
-                ProcessOutResult.Failure(it.code, it.message)
-            )
-        }
+        result
+            .onSuccess { delegateCache.remove()?.complete(uri = it) }
+            .onFailure { delegateCache.remove()?.complete(ProcessOutResult.Failure(it.code, it.message)) }
     }
 }
