+++ sdk/src/main/kotlin/com/processout/sdk/ui/web/ActivityResultDispatcher.kt
@@ -1,0 +1,9 @@
+package com.processout.sdk.ui.web
+
+import android.os.Parcelable
+import com.processout.sdk.core.ProcessOutActivityResult
+
+internal interface ActivityResultDispatcher<T : Parcelable> {
+
+    fun dispatch(result: ProcessOutActivityResult<T>)
+}
