+++ sdk/src/main/kotlin/com/processout/sdk/ui/web/WebAuthorizationDelegateCache.kt
@@ -7,6 +7,4 @@ internal interface WebAuthorizationDelegateCache {
     fun isCached(): Boolean
 
     fun remove(): WebAuthorizationDelegate?
-
-    fun clear()
 }
