--- sdk/src/main/kotlin/com/processout/sdk/ui/web/WebAuthorizationDelegateMemoryCache.kt
@@ -1,18 +1,0 @@
-package com.processout.sdk.ui.web
-
-internal object WebAuthorizationDelegateMemoryCache : WebAuthorizationDelegateCache {
-
-    override var delegate: WebAuthorizationDelegate? = null
-
-    override fun isCached() = delegate != null
-
-    override fun remove(): WebAuthorizationDelegate? {
-        val cached = delegate
-        clear()
-        return cached
-    }
-
-    override fun clear() {
-        delegate = null
-    }
-}
